Drop support for Excel 2013 (#66)
With version 1.1.0.0 we are introducing 2 new add-in commands; Upload
Insights and Import data. Since Excel 2013 does not support add-in
commands users on this platform would not be able to make use of these
new features.

Ref:
https://docs.microsoft.com/en-us/office/dev/add-ins/develop/create-addin-commands#step-2-create-a-task-pane-add-in

The choice to restrict it to ExcelApi 1.4 and above is informed by our
use of namedItemCollectionObject.add which is required to allow for
sheet bindings. This function was introduced in version 1.4 of the
ExcelApi.

Ref:
https://dev.office.com/reference/add-ins/excel/nameditemcollection